refactor naming and legend location
